refactored ClusterMarker and DebugOverlay components; added clsx and tailwind-merge dependencies; removed IconMarker and SearchBar components
