+++ package.json
@@ -19,6 +19,7 @@
 		"@types/mapbox-gl": "^3.4.1",
 		"@vis.gl/react-google-maps": "^1.4.2",
 		"axios": "^1.7.9",
+		"clsx": "^2.1.1",
 		"geojson": "^0.5.0",
 		"luxon": "^3.5.0",
 		"mapbox-gl": "^3.9.1",
@@ -30,6 +31,7 @@
 		"react-spinners": "^0.15.0",
 		"serve": "^14.2.4",
 		"supercluster": "^8.0.1",
+		"tailwind-merge": "^3.4.0",
 		"tailwindcss-safe-area": "^0.6.0",
 		"use-debounce": "^10.0.4",
 		"zustand": "^5.0.2"
