+++ src/components/BusinessInfoWindow/YelpLogo.tsx
@@ -1,4 +1,4 @@
-import yelpLogo from "../assets/yelp_logo_dark_bg.png";
+import yelpLogo from "../../assets/yelp_logo_dark_bg.png";
 
 export const YelpLogo = ({ alias }: { alias: string }) => (
 	<div className="shrink-0">
