+++ src/components/SearchBar/AiSearchButton.tsx
@@ -1,0 +1,13 @@
+import { motion } from 'motion/react';
+import { FaMagnifyingGlass } from 'react-icons/fa6';
+
+export const AiSearchButton = () => {
+	return (
+		<motion.button
+			whileHover={{ scale: 1.02 }}
+			className="w-[36px] h-[36px] flex justify-center items-center transition-all bg-gray-50/95 rounded-full outline-none hover:outline-none focus:outline-none p-0"
+		>
+			<FaMagnifyingGlass size={16} color={'black'} />
+		</motion.button>
+	);
+};
