+++ src/components/SearchBar/contants.ts
@@ -1,0 +1,8 @@
+import type { FilterColor } from '../../types/searchFilter';
+
+export const textColorVariants: Record<FilterColor, string> = {
+	green: 'text-green-600',
+	red: 'text-red-600',
+	purple: 'text-purple-600',
+	blue: 'text-blue-600',
+};
