+++ src/utils/cn.ts
@@ -1,0 +1,11 @@
+import { type ClassValue, clsx } from "clsx";
+import { twMerge } from "tailwind-merge";
+
+/**
+ * Combines clsx and tailwind-merge for optimal class merging.
+ * @param inputs - Class names or conditional class values.
+ * @returns A single string with merged class names.
+ */
+export const cn = (...inputs: ClassValue[]): string => {
+  return twMerge(clsx(inputs));
+};
